# Sort the conditional statements